qty changed to double properly working

+++ lib/features/cart/controllers/cart_controller.dart
@@ -52,7 +52,7 @@ class CartController extends GetxController {
   Future<void> addOrUpdateCart({
     required String pCode,
     required String iCode,
-    required int qty,
+    required double qty,
     required double rate,
   }) async {
     isLoading.value = true;
